Use SparkPlan transformer for physical pre/post-project
