+++ gluten-core/src/main/scala/io/glutenproject/execution/HashAggregateExecBaseTransformer.scala
@@ -62,17 +62,6 @@ abstract class HashAggregateExecBaseTransformer(
   @transient override lazy val metrics =
     BackendsApiManager.getMetricsApiInstance.genHashAggregateTransformerMetrics(sparkContext)
 
-  // The direct outputs of Aggregation.
-  lazy val allAggregateResultAttributes: List[Attribute] = {
-    val groupingAttributes = groupingExpressions.map(
-      expr => {
-        ConverterUtils.getAttrFromExpr(expr).toAttribute
-      })
-    groupingAttributes.toList ::: getAttrForAggregateExprs(
-      aggregateExpressions,
-      aggregateAttributes)
-  }
-
   protected def isCapableForStreamingAggregation: Boolean = {
     if (!conf.getConf(GlutenConfig.COLUMNAR_PREFER_STREAMING_AGGREGATE)) {
       return false
@@ -142,6 +131,13 @@ abstract class HashAggregateExecBaseTransformer(
         "Found not supported data type in group expression," +
           s"${groupingExpressions.map(_.dataType)}")
     }
+    aggregateExpressions.foreach {
+      expr =>
+        if (!checkAggFuncModeSupport(expr.aggregateFunction, expr.mode)) {
+          throw new UnsupportedOperationException(
+            s"Unsupported aggregate mode: ${expr.mode} for ${expr.aggregateFunction.prettyName}")
+        }
+    }
     doNativeValidation(substraitContext, relNode)
   }
 
@@ -157,25 +153,6 @@ abstract class HashAggregateExecBaseTransformer(
   // Members declared in org.apache.spark.sql.execution.AliasAwareOutputPartitioning
   override protected def outputExpressions: Seq[NamedExpression] = resultExpressions
 
-  // Check if Post-Projection is needed after the Aggregation.
-  def needsPostProjection: Boolean = {
-    // If the result expressions has different size with output attribute,
-    // post-projection is needed.
-    resultExpressions.size != allAggregateResultAttributes.size ||
-    // Compare each item in result expressions and output attributes. Attribute in Alias
-    // should be trimmed before checking.
-    resultExpressions.map(trimAliases).zip(allAggregateResultAttributes).exists {
-      case (exprAttr: Attribute, resAttr) =>
-        // If the result attribute and result expression has different name or type,
-        // post-projection is needed.
-        exprAttr.name != resAttr.name || exprAttr.dataType != resAttr.dataType
-      case _ =>
-        // If result expression is not instance of Attribute,
-        // post-projection is needed.
-        true
-    }
-  }
-
   protected def addFunctionNode(
       args: java.lang.Object,
       aggregateFunction: AggregateFunction,
@@ -189,65 +166,6 @@ abstract class HashAggregateExecBaseTransformer(
         modeToKeyWord(aggregateMode),
         ConverterUtils.getTypeNode(aggregateFunction.dataType, aggregateFunction.nullable)
       ))
-  }
-
-  /** This method calculates the output attributes of Aggregation. */
-  protected def getAttrForAggregateExprs(
-      aggregateExpressions: Seq[AggregateExpression],
-      aggregateAttributeList: Seq[Attribute]): List[Attribute] = {
-    val aggregateAttr = new ListBuffer[Attribute]()
-    val size = aggregateExpressions.size
-    var resIndex = 0
-    for (expIdx <- 0 until size) {
-      val exp: AggregateExpression = aggregateExpressions(expIdx)
-      resIndex = getAttrForAggregateExpr(exp, aggregateAttributeList, aggregateAttr, resIndex)
-    }
-    aggregateAttr.toList
-  }
-
-  protected def getAttrForAggregateExpr(
-      exp: AggregateExpression,
-      aggregateAttributeList: Seq[Attribute],
-      aggregateAttr: ListBuffer[Attribute],
-      index: Int): Int = {
-    var resIndex = index
-    val mode = exp.mode
-    val aggregateFunc = exp.aggregateFunction
-    // First handle the custom aggregate functions
-    if (
-      ExpressionMappings.expressionExtensionTransformer.extensionExpressionsMapping.contains(
-        aggregateFunc.getClass)
-    ) {
-      ExpressionMappings.expressionExtensionTransformer
-        .getAttrsIndexForExtensionAggregateExpr(
-          aggregateFunc,
-          mode,
-          exp,
-          aggregateAttributeList,
-          aggregateAttr,
-          index)
-    } else {
-      if (!checkAggFuncModeSupport(aggregateFunc, mode)) {
-        throw new UnsupportedOperationException(
-          s"Unsupported aggregate mode: $mode for ${aggregateFunc.prettyName}")
-      }
-      mode match {
-        case Partial | PartialMerge =>
-          val aggBufferAttr = aggregateFunc.inputAggBufferAttributes
-          for (index <- aggBufferAttr.indices) {
-            val attr = ConverterUtils.getAttrFromExpr(aggBufferAttr(index))
-            aggregateAttr += attr
-          }
-          resIndex += aggBufferAttr.size
-          resIndex
-        case Final =>
-          aggregateAttr += aggregateAttributeList(resIndex)
-          resIndex += 1
-          resIndex
-        case other =>
-          throw new UnsupportedOperationException(s"Unsupported aggregate mode: $other.")
-      }
-    }
   }
 
   protected def checkAggFuncModeSupport(
@@ -382,19 +300,66 @@ abstract class HashAggregateExecBaseTransformer(
       aggParams: AggregationParams,
       input: RelNode = null,
       validation: Boolean = false): RelNode
+}
 
-  /**
-   * Abstract classes do not support the use of the copy method, but case classes that inherit from
-   * the abstract class can directly use the copy method generated by Scala. Here, an interface is
-   * defined in the abstract class to call the copy method of the child case class.
-   */
-  def copySelf(
-      requiredChildDistributionExpressions: Option[Seq[Expression]] =
-        requiredChildDistributionExpressions,
-      groupingExpressions: Seq[NamedExpression] = groupingExpressions,
-      aggregateExpressions: Seq[AggregateExpression] = aggregateExpressions,
-      aggregateAttributes: Seq[Attribute] = aggregateAttributes,
-      initialInputBufferOffset: Int = initialInputBufferOffset,
-      resultExpressions: Seq[NamedExpression] = resultExpressions,
-      child: SparkPlan = child): HashAggregateExecBaseTransformer
+case class HashAggregateExecPullOutHelper(agg: BaseAggregateExec) extends AliasHelper {
+  // The direct outputs of Aggregation.
+  lazy val allAggregateResultAttributes: List[Attribute] =
+    agg.groupingExpressions.map(ConverterUtils.getAttrFromExpr(_)).toList :::
+      getAttrForAggregateExprs(agg.aggregateExpressions, agg.aggregateAttributes)
+
+  /** This method calculates the output attributes of Aggregation. */
+  protected def getAttrForAggregateExprs(
+      aggregateExpressions: Seq[AggregateExpression],
+      aggregateAttributeList: Seq[Attribute]): List[Attribute] = {
+    val aggregateAttr = new ListBuffer[Attribute]()
+    val size = aggregateExpressions.size
+    var resIndex = 0
+    for (expIdx <- 0 until size) {
+      val exp: AggregateExpression = aggregateExpressions(expIdx)
+      resIndex = getAttrForAggregateExpr(exp, aggregateAttributeList, aggregateAttr, resIndex)
+    }
+    aggregateAttr.toList
+  }
+
+  protected def getAttrForAggregateExpr(
+      exp: AggregateExpression,
+      aggregateAttributeList: Seq[Attribute],
+      aggregateAttr: ListBuffer[Attribute],
+      index: Int): Int = {
+    var resIndex = index
+    val mode = exp.mode
+    val aggregateFunc = exp.aggregateFunction
+    // First handle the custom aggregate functions
+    if (
+      ExpressionMappings.expressionExtensionTransformer.extensionExpressionsMapping.contains(
+        aggregateFunc.getClass)
+    ) {
+      ExpressionMappings.expressionExtensionTransformer
+        .getAttrsIndexForExtensionAggregateExpr(
+          aggregateFunc,
+          mode,
+          exp,
+          aggregateAttributeList,
+          aggregateAttr,
+          index)
+    } else {
+      mode match {
+        case Partial | PartialMerge =>
+          val aggBufferAttr = aggregateFunc.inputAggBufferAttributes
+          for (index <- aggBufferAttr.indices) {
+            val attr = ConverterUtils.getAttrFromExpr(aggBufferAttr(index))
+            aggregateAttr += attr
+          }
+          resIndex += aggBufferAttr.size
+          resIndex
+        case Final =>
+          aggregateAttr += aggregateAttributeList(resIndex)
+          resIndex += 1
+          resIndex
+        case other =>
+          throw new UnsupportedOperationException(s"Unsupported aggregate mode: $other.")
+      }
+    }
+  }
 }
