+++ gluten-core/src/main/scala/io/glutenproject/execution/SortExecTransformer.scala
@@ -51,23 +51,33 @@ case class SortExecTransformer(
   override def output: Seq[Attribute] = child.output
 
   override def outputOrdering: Seq[SortOrder] = child match {
-    case project: ProjectExecTransformer if project.isPreProject =>
-      // After adding a pre-project to the sort, the expressions in sortOrder are replaced
-      // with attributes. To ensure that the downstream post-project can obtain the correct
-      // outputOrdering, it is necessary to restore the current sortOrder based on the
-      // pre-project's alias. Otherwise, when verifying whether the output attributes of
-      // the downstream are a subset of the references of sortOrder.child, it will not pass.
-      // Please check AliasAwareQueryOutputOrdering.outputOrdering in Spark-3.4.
-      val aliasMap = getAliasMap(project.projectList)
-      sortOrder.map(_.mapChildren(replaceAlias(_, aliasMap)).asInstanceOf[SortOrder])
-    case _ =>
-      sortOrder
+    case project: ProjectExecTransformer =>
+      getSortOrderFromPreProject(project.projectList)
+    case InputIteratorTransformer(InputAdapter(r2c: RowToColumnarExecBase)) =>
+      r2c.child match {
+        case project: ProjectExec =>
+          // This case means pre-project is fallback.
+          getSortOrderFromPreProject(project.projectList)
+        case _ => sortOrder
+      }
+    case _ => sortOrder
   }
 
   override def outputPartitioning: Partitioning = child.outputPartitioning
 
   override def requiredChildDistribution: Seq[Distribution] =
     if (global) OrderedDistribution(sortOrder) :: Nil else UnspecifiedDistribution :: Nil
+
+  private def getSortOrderFromPreProject(projectList: Seq[NamedExpression]): Seq[SortOrder] = {
+    // After adding a pre-project to the sort, the expressions in sortOrder are replaced
+    // with attributes. To ensure that the downstream post-project can obtain the correct
+    // outputOrdering, it is necessary to restore the current sortOrder based on the
+    // pre-project's alias. Otherwise, when verifying whether the output attributes of
+    // the downstream are a subset of the references of sortOrder.child, it will not pass.
+    // Please check AliasAwareQueryOutputOrdering.outputOrdering in Spark-3.4.
+    val aliasMap = getAliasMap(projectList)
+    sortOrder.map(_.mapChildren(replaceAlias(_, aliasMap)).asInstanceOf[SortOrder])
+  }
 
   def getRelNode(
       context: SubstraitContext,
