+++ gluten-ut/spark34/src/test/scala/org/apache/spark/sql/extension/GlutenCollapseProjectExecTransformerSuite.scala
@@ -108,7 +108,7 @@ class GlutenCollapseProjectExecTransformerSuite extends GlutenSQLTestsTrait {
             )
             assert(
               getExecutedPlan(df).exists {
-                case _ @ProjectExecTransformer(_, _: ProjectExecTransformer, _) => true
+                case _ @ProjectExecTransformer(_, _: ProjectExecTransformer) => true
                 case _ => false
               } == !collapsed
             )
